Fix issues introduced in the bitwiseness work and align semantics; fix a few pre-existing bugs.

+++ crates/infotheory/benches/mcts_planners.rs
@@ -96,10 +96,10 @@ impl AgentSimulator for BenchAgent {
         let mix = self.burn(self.step_work.max(1));
         if self.emit_reward {
             self.emit_reward = false;
-            ((self.last_action ^ mix) & 1) as u64
+            (self.last_action ^ mix) & 1
         } else {
             self.emit_reward = true;
-            ((mix >> 5) & 1) as u64
+            (mix >> 5) & 1
         }
     }
